v2_02_76_src
Former-commit-id: e638d38fb9407ba82b1423e5596ccadd84aaecd3

+++ host/dib.cpp
@@ -319,7 +319,7 @@ DWORD CDib::Save(CFile& file) const
 DWORD CDib::Read(CFile& file, BOOL bFromResource)
 {
 	DWORD dwReadBytes = 0;
-	DWORD dwLength = file.GetLength();
+	DWORD dwLength = (DWORD)file.GetLength();
 	
 	// Ensures no memory leaks will occur
 	Free();
